% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lazega_analysis.R
\name{comp_dyad_inner_prod}
\alias{comp_dyad_inner_prod}
\title{Compute the inner product of each activated dyad.}
\usage{
comp_dyad_inner_prod(data = Lazega_lawyer_network)
}
\arguments{
\item{data}{}
}
\value{
a list of inner products
}
\description{
Compute the inner product of each activated dyad.
}
