% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lazega_analysis.R
\name{comp_indep_test}
\alias{comp_indep_test}
\title{Compute the independence test statistic}
\usage{
comp_indep_test(data = Lazega_lawyer_network, B = 100)
}
\arguments{
\item{data}{}
}
\value{
a list of inner products
}
\description{
Construct the test statistic by computing each dyad's inner product of its neighboring dyads
}
