% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lazega_analysis.R
\name{comp_indep_test_dyad}
\alias{comp_indep_test_dyad}
\title{Compute the independence test statistic for each activated dyad}
\usage{
comp_indep_test_dyad(data = Lazega_lawyer_network, B = 100)
}
\arguments{
\item{data}{}
}
\value{
a list of inner products. Each element in the list corresponds to a dyad.
 The first element of each list is the inner product of all neighboring dyads, and the rest are inner products after random permutation.
}
\description{
Construct the test statistic by computing each dyad's inner product of its neighboring dyads
}
