% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{comp_indep_test_network}
\alias{comp_indep_test_network}
\title{Compute the independence test statistic for the whole network}
\usage{
comp_indep_test_network(data = Lazega_lawyer_network, B = 100)
}
\arguments{
\item{data}{n by 3 matrix with each column indicating index1, index2, and layer. The dyad index must be lexicographically ordered.}

\item{B}{}
}
\description{
Construct the network-wise test statistic by summing over each dyad's inner product of its neighboring dyads
}
