% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{comp_inner_prod_avg}
\alias{comp_inner_prod_avg}
\title{Compute the average neighboring and non-neighboring inner product of each activated dyad.}
\usage{
comp_inner_prod_avg(data = Lazega_lawyer_network)
}
\arguments{
\item{data}{}
}
\description{
This function computes the neighboring and non-neighboring inner product mean of each activated dyad.
}
