% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lazega_analysis.R
\name{compute_suffstats}
\alias{compute_suffstats}
\title{Title
Only calculate sufficient stats for models without interactions}
\usage{
compute_suffstats(adj, H = 1)
}
\arguments{
\item{adj}{}

\item{H}{}
}
\description{
Title
Only calculate sufficient stats for models without interactions
}
