% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lazega_analysis.R
\name{compute_suffstats_Lazega}
\alias{compute_suffstats_Lazega}
\title{Title}
\usage{
compute_suffstats_Lazega(net_data = Lazega_lawyer_network, N = 71)
}
\arguments{
\item{net_data}{}

\item{N}{}
}
\description{
Title
}
