% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lazega_analysis.R
\name{draw_box_plot_Lazega}
\alias{draw_box_plot_Lazega}
\title{Title}
\usage{
draw_box_plot_Lazega(
  reproduced_suff,
  obs,
  dimx_name = c("Coworker", "Advice", "Friendship", "CxA", "CxF", "AxF")
)
}
\arguments{
\item{reproduced_suff}{}

\item{obs}{}

\item{dimx_name}{}
}
\description{
Title
}
