% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_net.R
\name{er.simulate}
\alias{er.simulate}
\title{Title}
\usage{
er.simulate(
  theta,
  N,
  samp_num = 1,
  burnin = 100,
  k = 3,
  mdim = 6,
  mterm = rep("ml_order2", mdim),
  intv = 3,
  seed = 0,
  gy = 1
)
}
\arguments{
\item{theta}{model-generating parameter by lexicographic order}

\item{N}{number of nodes}

\item{samp_num}{number of simulated samples}

\item{burnin}{number of burn-ins for the MCMC algorithm}

\item{k}{number of layers}

\item{mdim}{number of parameter dimensions}

\item{mterm}{model terms}

\item{intv}{length of the update interval for the MCMC algorithm}

\item{seed}{random seed}

\item{gy}{edge probability for the Bernoulli basis network}
}
\value{
return a list of k-layer network with a Bernoulli basis network and ...
}
\description{
Title
}
