% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_net.R
\name{est_ml}
\alias{est_ml}
\title{Title}
\usage{
est_ml(
  NetMat,
  N,
  samp_num = 1,
  burnin = 100,
  k = 3,
  H = 2,
  mdim,
  mterm = "BER",
  intv = 3,
  seed = 0,
  iter_max = 30000,
  basis_arguments = c(1, 0)
)
}
\arguments{
\item{N}{}

\item{samp_num}{}

\item{burnin}{}

\item{k}{}

\item{H}{}

\item{mdim}{}

\item{mterm}{}

\item{intv}{}

\item{seed}{}

\item{iter_max}{}

\item{basis_arguments}{}

\item{theta}{}
}
\description{
Title
}
