% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_basis.R
\name{gen_sbm}
\alias{gen_sbm}
\title{Title}
\usage{
gen_sbm(N, node_membs, p_within, p_between)
}
\arguments{
\item{N}{}

\item{node_membs}{}

\item{p_within}{}

\item{p_between}{}
}
\description{
Title
}
