% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_basis.R
\name{generate_SBM_basis}
\alias{generate_SBM_basis}
\title{Title}
\usage{
generate_SBM_basis(
  N = 3 * 10,
  num_of_block = 3,
  p_within = 0.5,
  p_between = 0.1,
  M = 0
)
}
\arguments{
\item{N}{}

\item{num_of_block}{}

\item{p_within}{}

\item{p_between}{}

\item{M}{}
}
\description{
Title
}
