% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{link_tracing_samp}
\alias{link_tracing_samp}
\title{Title}
\usage{
link_tracing_samp(data = Lazega_lawyer_network, init = 3, w = 2, N = 0)
}
\arguments{
\item{data}{}

\item{init}{number of initial nodes}

\item{w}{number of waves}

\item{N}{}
}
\description{
Title
}
