\name{mlayernetwork-package}
\alias{mlayernetwork-package}
\alias{mlayernetwork}
\docType{package}
\title{
\packageTitle{mlayernetwork}
}
\description{
\packageDescription{mlayernetwork}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mlayernetwork}
\packageIndices{mlayernetwork}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mlayernetwork}

Maintainer: \packageMaintainer{mlayernetwork}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
