% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_compute_dyad_suffstats}
\alias{rcpp_compute_dyad_suffstats}
\title{rcpp_compute_dyad_suffstats}
\usage{
rcpp_compute_dyad_suffstats(
  RNETWORK,
  rmodel_dim,
  rnum_nodes,
  rnum_layers,
  rhighest_order
)
}
\description{
Compute dyadic sufficient statistics for multilayer networks with up to H-way interactions
}
