% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_estimate_model_ml}
\alias{rcpp_estimate_model_ml}
\title{rcpp_estimate_model_ml}
\usage{
rcpp_estimate_model_ml(
  RNETWORK,
  rIter_max,
  rsamp_num,
  rburnin,
  rinterval,
  rmodel_dim,
  model_terms,
  rnum_nodes,
  rnum_layers,
  rcheck_chull,
  random_seeds,
  g
)
}
\description{
Estimate (MPLE) parameters for network-separable multilayer network models with 2-way interactions
}
