% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_estimate_model_ml_3way}
\alias{rcpp_estimate_model_ml_3way}
\title{rcpp_estimate_model_ml_3way}
\usage{
rcpp_estimate_model_ml_3way(
  RNETWORK,
  rNR_tol,
  rNR_max,
  rMCMLE_max,
  rsamp_num,
  rburnin,
  rinterval,
  rmodel_dim,
  model_terms,
  rnum_nodes,
  rnum_layers,
  rcheck_chull,
  random_seeds,
  g
)
}
\description{
Estimate (MPLE) parameters for network-separable multilayer network models with 3-way interactions
}
