% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_estimate_model_ml_Hway}
\alias{rcpp_estimate_model_ml_Hway}
\title{rcpp_estimate_model_ml_Hway}
\usage{
rcpp_estimate_model_ml_Hway(
  RNETWORK,
  rsamp_num,
  rburnin,
  rinterval,
  rmodel_dim,
  model_terms,
  rnum_nodes,
  rnum_layers,
  random_seeds,
  g
)
}
\description{
Estimate (MPLE) parameters for network-separable multilayer network models with up to H-way interactions
}
