% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_exact_simulate_ml}
\alias{rcpp_exact_simulate_ml}
\title{rcpp_exact_simulate_ml}
\usage{
rcpp_exact_simulate_ml(
  rsamp_num,
  rburnin,
  rinterval,
  rmodel_dim,
  model_terms,
  rnum_nodes,
  rnum_layers,
  rtheta,
  rand_seed
)
}
\description{
Simulate a multilayer network model by its joint density
}
\examples{
rcpp_exact_simulate_ml(1,1,3,7,rep("ml_order2",7),10,3,c(0,0.693,0.693,0,0.693,0,0),23333)

}
