% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_simulate_ml}
\alias{rcpp_simulate_ml}
\alias{rcpp_simulate_ml_LSM}
\alias{rcpp_simulate_ml_SBM}
\title{rcpp_simulate_ml}
\usage{
rcpp_simulate_ml_LSM(
  rsamp_num,
  rburnin,
  rinterval,
  rmodel_dim,
  model_terms,
  rnum_nodes,
  rnum_layers,
  rtheta,
  rand_seed,
  g,
  fe,
  nmu,
  nsd
)

rcpp_simulate_ml_SBM(
  rsamp_num,
  rburnin,
  rinterval,
  rmodel_dim,
  model_terms,
  rnum_nodes,
  rnum_layers,
  rtheta,
  rand_seed,
  g,
  block,
  p1,
  p2
)

rcpp_simulate_ml(
  rsamp_num,
  rburnin,
  rinterval,
  rmodel_dim,
  model_terms,
  rnum_nodes,
  rnum_layers,
  rtheta,
  rand_seed,
  g
)
}
\description{
Simulate a multilayer network model

Simulate a multilayer network model

Simulate a multilayer network model
}
\examples{
rcpp_simulate_ml_SBM(1,100,1,2,c("withinlayer","crosslayer"),3,2,c(0.4,0.7))

rcpp_simulate_ml_SBM(1,100,1,2,c("withinlayer","crosslayer"),3,2,c(0.4,0.7))

}
