% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_simulate_ml_Hway}
\alias{rcpp_simulate_ml_Hway}
\title{rcpp_simulate_ml_Hway}
\usage{
rcpp_simulate_ml_Hway(
  rsamp_num,
  rburnin,
  rinterval,
  rmodel_dim,
  model_terms,
  rnum_nodes,
  rnum_layers,
  rtheta,
  rhighest_order,
  rand_seed,
  arguments
)
}
\description{
Simulate a network-separable multilayer network model with arbitrary H-way interactions
}
