% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_simulate_ml_LSM}
\alias{rcpp_simulate_ml_LSM}
\title{rcpp_simulate_ml_LSM}
\usage{
rcpp_simulate_ml_LSM(
  rsamp_num,
  rburnin,
  rinterval,
  rmodel_dim,
  model_terms,
  rnum_nodes,
  rnum_layers,
  rtheta,
  rand_seed,
  g,
  fe,
  nmu,
  nsd
)
}
\description{
Simulate a network-separable multilayer network model with basis network generated by latent space model
}
