% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_simulate_ml_SBM}
\alias{rcpp_simulate_ml_SBM}
\title{rcpp_simulate_ml_SBM}
\usage{
rcpp_simulate_ml_SBM(
  rsamp_num,
  rburnin,
  rinterval,
  rmodel_dim,
  model_terms,
  rnum_nodes,
  rnum_layers,
  rtheta,
  rand_seed,
  g,
  block,
  p1,
  p2
)
}
\description{
Simulate a network-separable multilayer network model with basis network generated by stochastic block model
}
