% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lazega_analysis.R
\name{reproduce_Lazega}
\alias{reproduce_Lazega}
\title{Title}
\usage{
reproduce_Lazega(
  m = 10,
  theta = c(-1.4497737, -3.3336582, -2.6945731, 1.8011888, 0.2176545, 2.4582079)
)
}
\arguments{
\item{m}{}

\item{theta}{}
}
\description{
Title
}
