% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_net.R
\name{samp_ml}
\alias{samp_ml}
\title{Title}
\usage{
samp_ml(
  theta,
  N = 10,
  samp_num = 1,
  burnin = 100,
  k = 3,
  mdim,
  mterm = "BER",
  intv = 3,
  H = 2,
  seed = 0,
  basis_arguments = c(0.5, 0)
)
}
\arguments{
\item{theta}{model-generating parameter vector by lexicographic order}

\item{N}{number of nodes}

\item{samp_num}{number of simulated samples}

\item{burnin}{number of burn-ins for the MCMC algorithm}

\item{k}{number of layers}

\item{mdim}{number of parameter dimensions}

\item{mterm}{model terms}

\item{intv}{length of the update interval for the MCMC algorithm}

\item{seed}{random seed}

\item{gy}{edge probability for the Bernoulli basis network}
}
\value{
return a list of k-layer network with a Bernoulli basis network and ...
}
\description{
Title
}
