% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_net.R
\name{sim_est}
\alias{sim_est}
\title{Title}
\usage{
sim_est(theta, N, samp_num, burnin, k, mdim, mterm, intv, iter_max, seed, gy)
}
\arguments{
\item{theta}{model-generating parameter}

\item{N}{network size}

\item{samp_num}{}

\item{burnin}{}

\item{k}{}

\item{mdim}{}

\item{mterm}{}

\item{intv}{}

\item{iter_max}{}

\item{seed}{}

\item{gy}{}
}
\value{
a list
}
\description{
Title
}
