% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_net.R
\name{sim_est_wrapper}
\alias{sim_est_wrapper}
\title{Title}
\usage{
sim_est_wrapper(
  sim_id,
  thetas,
  N,
  samp_num = 1,
  burnin = 100,
  k,
  H,
  mdim,
  mterm,
  intv = 3,
  seeds,
  iter_max = 30000,
  basis_arguments
)
}
\arguments{
\item{N}{}

\item{samp_num}{}

\item{burnin}{}

\item{k}{}

\item{H}{}

\item{mdim}{}

\item{mterm}{}

\item{intv}{}

\item{iter_max}{}

\item{basis_arguments}{}

\item{theta}{}

\item{seed}{}
}
\description{
Title
}
