% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_simulate_ml}
\alias{rcpp_simulate_ml}
\title{rcpp_simulate_ml}
\usage{
rcpp_simulate_ml(
  rsamp_num,
  rburnin,
  rinterval,
  rmodel_dim,
  model_terms,
  rnum_nodes,
  rnum_layers,
  rtheta,
  rand_seed,
  g
)
}
\description{
Simulate a network-separable multilayer network model with 2-way interactions
}
