% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_net.R
\name{sim_est}
\alias{sim_est}
\title{Title}
\usage{
sim_est(
  theta,
  N,
  samp_num = 1,
  burnin = 100,
  k = 3,
  H = 2,
  mdim,
  mterm = "BER",
  intv = 3,
  seed = 0,
  iter_max = 30000,
  basis_arguments = c(1, 0)
)
}
\arguments{
\item{theta}{model-generating parameter}

\item{N}{network size}

\item{samp_num}{}

\item{burnin}{}

\item{k}{}

\item{mdim}{}

\item{mterm}{}

\item{intv}{}

\item{seed}{}

\item{iter_max}{}

\item{gy}{}
}
\value{
a list
}
\description{
Title
}
